pyproject.toml
[project]
name = "collective"
version = "0.1.0"
description = "A structured collective dialogue platform"
requires-python = ">=3.14"
dependencies = [
    "Django>=6.0,<7.0",
    "psycopg[binary]>=3.3",
    "django-allauth>=65.16",
    "django-treebeard>=5.0",
    "whitenoise>=6.12",
    "python-dotenv>=1.2",
    "django-tinymce>=5.0",
    "nh3>=0.3",
    "Markdown>=3.10",
    "filetype>=1.2",
    "python-docx>=1.2",
]

[tool.setuptools.packages.find]
include = ["collective*", "apps*"]

[project.optional-dependencies]
dev = [
    "ruff>=0.15",
    "mypy>=1.20",
    "django-stubs>=6.0",
    "pytest>=9.0",
    "pytest-django>=4.12",
    "pytest-cov>=7.0",
    "factory-boy>=3.3",
    "debugpy>=1.8",
]

[tool.ruff]
target-version = "py314"
line-length = 120
src = ["collective"]
# Factory modules stay linted; only generated migrations are excluded.
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "DJ"]

[tool.mypy]
python_version = "3.14"
plugins = ["mypy_django_plugin.main"]
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[[tool.mypy.overrides]]
module = "treebeard.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["apps.*.tests.*", "conftest"]
ignore_errors = true

[[tool.mypy.overrides]]
module = "apps.*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "collective.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "collective.settings.test"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
